Erika OO, MISRA check: Flexelint Error 960/ MISRA MISRA 10.1.
This patch fixes the following MISRA/Flexelint error for the
semaphores modules:

Violates MISRA 2004 Required Rule 10.1, Implicit conversion
changes signedness

git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@1991 88d5ba13-baa4-4915-ae7a-41304cef7dd7

+++ pkg/kernel/oo/src/ee_semgetvalue.c
@@ -57,13 +57,13 @@ int EE_oo_GetValueSem(SemRefType Sem)
 
 #if defined(__OO_ECC1__) || defined(__OO_ECC2__)
   if (Sem->first == EE_NIL) {
-    returnvalue = Sem->count;
+    returnvalue = (int)(Sem->count);
   }
   else {
     returnvalue = -1;
   }
 #else
-  returnvalue = Sem->count;
+  returnvalue = (int)(Sem->count);
 #endif
 
   EE_hal_end_nested_primitive(flag);
